refactor: URLSession dataTask 수정 URLSession dataTask 부분 공용으로 쓸 수 있도록 메서드 추가 NetworkManager 에 dataTask 공용 메서드 사용 NetworkProtocol.swift -> URLSessionProtocol.swift 로 네이밍 변경
