+++ OpenMarket/OpenMarket/Controller/OpenMarketViewController.swift
@@ -40,3 +40,7 @@ extension OpenMarketViewController: UICollectionViewDataSource {
 extension OpenMarketViewController: UICollectionViewDelegate {
     
 }
+
+extension OpenMarketViewController {
+    
+}
